made float not double

+++ tell/models/BMRelModel.py
@@ -127,7 +127,7 @@ class BMRelModel(Model):
         # TODO: check if we want to address sentence length and calculate mean w.r.t that
         # h = torch.mean(hiddens, dim=2)  # [B, K, 1024]
         sums = torch.sum(hiddens, dim=2)
-        h = sums / np.tile(np.argmin(mask, axis=-1)[:, :, np.newaxis], sums.shape[-1])
+        h = sums / np.tile(np.argmin(mask, axis=-1)[:, :, np.newaxis], sums.shape[-1]).astype(np.float32)
         # nan -> 0 before passing through layers, then we mask these paragraphs out anyway
         # h = torch.nan_to_num(h) # doesn't exist in torch 1.5.1
         h[torch.isnan(h)] = 0
